Completion of v2.3.0
Main changes:
Allows users to chose an experiment if multiple exist in a given date (previously the code wouldn't run at all).
Allows users to now convert annular scans (they will be treated like linear scans).
Added CITATION.cff.
Changed file name formatting and removed v1.4.0.
