+++ convert_OCT_files.py
@@ -1,10 +1,10 @@
-#Open_OCT_file_2.2.0.py
-#Converting OCT files to tif files
-#Made by Brandon Anderson, University of Pennsylvania
+# convert_OCT_files.py
+# Converting OCT files to tif and avi files
+# Made by Brandon Anderson, University of Pennsylvania
 
-#For this macro to work, you need the following files:
-#OCT reader plugin - ImageJ macro
-#OCT averager plugin - ImageJ macro
+# For this macro to work, you need the following files:
+# OCT reader plugin - ImageJ macro
+# OCT averager plugin - ImageJ macro
 
 import os
 import psutil       # For opening ImageJ
@@ -15,12 +15,16 @@ import shutil       # For deleting directories
 import pygetwindow as gw    #for making ImageJ an active window
 import tkinter as tk    # For making dialog boxes
 import tkinter.messagebox as messagebox
-from tkinter import ttk, filedialog, messagebox
+from tkinter import ttk, filedialog, messagebox, simpledialog
 import inspect      # For error messages
+import ast          # For converting variables in text file to proper variable formats
+import sys          # For exiting dialog boxes
+import pyscreeze
+pyscreeze.USE_IMAGE_NOT_FOUND_EXCEPTION = False
 
 # You will need to make sure to download psutil, pyautogui, and pygetwindow as they don't come with base Python
 
-
+confidence = 0.95
 # For updates, look for # Update:
 # Update: volume scan
 # Update: change peripheral folder checkbox to entry box
@@ -63,7 +67,7 @@ def show_error_message(message, root=None):
 def wait_to_appear(image_file, max_wait_time):      # This looks for dialog boxes or buttons to appear, and won't let the script continue until it is found or if the time reaches the max wait time
     search_start_time = time.time()
     while time.time() - search_start_time < max_wait_time:
-        if pyautogui.locateOnScreen(image_file):
+        if pyautogui.locateOnScreen(image_file, confidence=confidence):
             time.sleep(0.5)
             return True
         time.sleep(0.5)
@@ -136,72 +140,37 @@ def retrieve_variables(preferences_variables, key):
     try:
         return preferences_variables[key]
     except KeyError:        # If these variables don't exist in the document, here are the defaults
-        if key == "image_location_list":
-            return ["central", "temporal", "nasal", "superior", "inferior"]
-        elif key == "od_before_os":
-            return True
-        elif key == "scan_modes":
-            return ['radial', 'linear', 'linear', 'radial', 'radial']
-        # Update: change peripheral folder checkbox to entry box
-        elif key == "subfolder_entry":
-            return [False, True, True, True, True]
-        # Update: change peripheral folder checkbox to entry box
-        elif key == "subfolder_name":
-            return "Peripheral images"
-        elif key == "horizontal_image_list":
-            return ["horizontal", "", "", "", ""]
-        elif key == "vertical_image_list":
-            return ["vertical", "", "", "superior", "inferior"]
-        elif key == "unaveraged_images":
-            return False
-        elif key == "min_contrast":
-            return "25"
-        elif key == "max_contrast":
-            return "215"
-        elif key == "crop_amount":
-            return "480"
-        
+        if key == "preset_options_dict":
+            return {"Previous settings": {"default_locations": ["horizontal, vertical", "temporal", "nasal", "delete, superior", "delete, inferior"], "default_scan_type": ["radial", "linear", "linear", "radial", "radial"], "default_subfolder": [False, True, True, True, True], "default_min_contrast": 25, "default_max_contrast": 215, "default_crop_amount": 480, "default_subfolder_name": "Peripheral images", "od_os_checkbox_boolean": True, "imagej_checkbox_boolean": False, "unaveraged_checkbox_boolean": False}}
 
 
-image_location_list = retrieve_variables(preferences_variables, 'image_location_list')
-od_before_os = retrieve_variables(preferences_variables, 'od_before_os')
-scan_modes = retrieve_variables(preferences_variables, 'scan_modes')
-subfolder_entry = retrieve_variables(preferences_variables, 'subfolder_entry')
-subfolder_name = retrieve_variables(preferences_variables, 'subfolder_name')
-horizontal_image_list = retrieve_variables(preferences_variables, 'horizontal_image_list')
-vertical_image_list = retrieve_variables(preferences_variables, 'vertical_image_list')
-unaveraged_images = retrieve_variables(preferences_variables, 'unaveraged_images')
-min_contrast = retrieve_variables(preferences_variables, 'min_contrast')
-max_contrast = retrieve_variables(preferences_variables, 'max_contrast')
-crop_amount = retrieve_variables(preferences_variables, "crop_amount")
+
+preset_options_dict = retrieve_variables(preferences_variables, 'preset_options_dict')
+if type(preset_options_dict) == str:
+    preset_options_dict = ast.literal_eval(preset_options_dict)
 
 
-# Converting str to whatever it actually is, since only strings come from the txt file
-if type(image_location_list) == str:
-    image_location_list = eval(image_location_list)
-if type(od_before_os) == str:
-    od_before_os = eval(od_before_os)
-if type(unaveraged_images) == str:
-    unaveraged_images = eval(unaveraged_images)
-if type(scan_modes) == str:
-    scan_modes = eval(scan_modes)
-if type(subfolder_entry) == str:
-    subfolder_entry = eval(subfolder_entry)
-if type(horizontal_image_list) == str:
-    horizontal_image_list = eval(horizontal_image_list)
-if type(vertical_image_list) == str:
-    vertical_image_list = eval(vertical_image_list)
+image_location_list = preset_options_dict["Previous settings"]["default_locations"]
+od_before_os = preset_options_dict["Previous settings"]["od_os_checkbox_boolean"]
+scan_modes = preset_options_dict["Previous settings"]["default_scan_type"]
+subfolder_entry = preset_options_dict["Previous settings"]["default_subfolder"]
+subfolder_name = preset_options_dict["Previous settings"]["default_subfolder_name"]
+unaveraged_images = preset_options_dict["Previous settings"]["unaveraged_checkbox_boolean"]
+min_contrast = preset_options_dict["Previous settings"]["default_min_contrast"]
+max_contrast = preset_options_dict["Previous settings"]["default_max_contrast"]
+crop_amount = preset_options_dict["Previous settings"]["default_crop_amount"]
+
 
 
 
 
 #This first dialog box checks to see what images were taken and in which order
-def location_input_dialog_box():
+def first_dialog_box():
     row_num = 1 # Starting row number for entry and checkbox widgets
 
     def add_entry_row(mode="linear", subfolder_mode=False):
         nonlocal row_num
-        entry_row = tk.Entry(dialog_frame)
+        entry_row = tk.Entry(dialog_frame, width=50)
         entry_row.grid(row=row_num, column=0, sticky='w')
         entry_rows.append(entry_row)
 
@@ -225,22 +194,11 @@ def location_input_dialog_box():
         checkbox_volumes.append(checkbox_volume)
 
 
-        # Adding the horizontal and vertical entry boxes
-        entry_horizontal = tk.Entry(dialog_frame)
-        entry_horizontal.grid(row=row_num, column=4, sticky='w')
-        entry_horizontals.append(entry_horizontal)
-
-        entry_vertical = tk.Entry(dialog_frame)
-        entry_vertical.grid(row=row_num, column=5, sticky='w')
-        entry_verticals.append(entry_vertical)
-        if mode != "radial":
-            entry_horizontal.config(state='disabled')
-            entry_vertical.config(state='disabled')
 
         # Adding subfolder checkboxes
         subfolder_var = tk.BooleanVar(value=subfolder_mode)
         checkbox_subfolder = tk.Checkbutton(dialog_frame, variable=subfolder_var)
-        checkbox_subfolder.grid(row=row_num, column=6, sticky='s')
+        checkbox_subfolder.grid(row=row_num, column=4, sticky='s')
         checkbox_subfolders.append(checkbox_subfolder)
         subfolder_vars.append(subfolder_var)
         # Update: change peripheral folder checkbox to entry box
@@ -268,12 +226,7 @@ def location_input_dialog_box():
             checkbox_volume = checkbox_volumes.pop()
             checkbox_volume.destroy()
 
-            # Remove the corresponding horizontal and vertical entry boxes from the lists and destroy them
-            entry_horizontal = entry_horizontals.pop()
-            entry_horizontal.destroy()
 
-            entry_vertical = entry_verticals.pop()
-            entry_vertical.destroy()
 
             # Update: change peripheral folder checkbox to entry box
             checkbox_subfolder = checkbox_subfolders.pop()
@@ -289,56 +242,46 @@ def location_input_dialog_box():
         linear_vars[index].set(True)
         radial_vars[index].set(False)
         volume_vars[index].set(False)
-        entry_horizontals[index].delete(0, tk.END)
-        entry_verticals[index].delete(0, tk.END)
-        entry_horizontals[index].config(state='disabled')
-        entry_verticals[index].config(state='disabled')
+
 
     def select_radial(index):
         linear_vars[index].set(False)
         radial_vars[index].set(True)
         volume_vars[index].set(False)
-        entry_horizontals[index].config(state='normal')
-        entry_verticals[index].config(state='normal')
 
     def select_volume(index):
         linear_vars[index].set(False)
         radial_vars[index].set(False)
         volume_vars[index].set(True)
-        entry_horizontals[index].delete(0, tk.END)
-        entry_verticals[index].delete(0, tk.END)
-        entry_horizontals[index].config(state='disabled')
-        entry_verticals[index].config(state='disabled')
 
 
-    def restore_defaults():
+    def set_new_preset(preset_option):
         # Update: custom defaults - need to define at function that sees user settings (much earlier in code)
-        default_values = ["central", "temporal", "nasal", "superior", "inferior"]
-        default_horizontal = ["horizontal", "", "", "", ""]
-        default_scan_type = ["radial", "linear", "linear", "radial", "radial"]
-        default_vertical = ["vertical", "", "", "superior", "inferior"]
-        default_subfolder = [False, True, True, True, True]
-        default_min_contrast = 25   # Update: other - need to define this at beginning
-        default_max_contrast = 215  # Update: other - need to define this at beginning
-        default_crop_amount = 480  # Update: other - need to define this at beginning
-        default_subfolder_name = "Peripheral images"
-        current_row_count = len(entry_rows)
-        od_os_checkbox_boolean = True
-        imagej_checkbox_boolean = False
-        unaveraged_checkbox_boolean = False
+        default_locations = preset_options_dict[preset_option]["default_locations"]
+        default_scan_type = preset_options_dict[preset_option]["default_scan_type"]
+        default_subfolder = preset_options_dict[preset_option]["default_subfolder"]
+        default_min_contrast = preset_options_dict[preset_option]["default_min_contrast"]
+        default_max_contrast = preset_options_dict[preset_option]["default_max_contrast"]
+        default_crop_amount = preset_options_dict[preset_option]["default_crop_amount"]
+        default_subfolder_name = preset_options_dict[preset_option]["default_subfolder_name"]
+        od_os_checkbox_boolean = preset_options_dict[preset_option]["od_os_checkbox_boolean"]
+        imagej_checkbox_boolean = preset_options_dict[preset_option]["imagej_checkbox_boolean"]
+        unaveraged_checkbox_boolean = preset_options_dict[preset_option]["unaveraged_checkbox_boolean"]
 
+        current_row_count = len(entry_rows)
+        
         # If the current number of rows is less than five, add new rows
-        while current_row_count < 5:
+        while current_row_count < len(default_locations):
             add_entry_row()
             current_row_count += 1
 
         # If the current number of rows is more than five, remove excess rows
-        while current_row_count > 5:
+        while current_row_count > len(default_locations):
             remove_entry_row()
             current_row_count -= 1
 
         # Update the content of the first column with default values
-        for i, default_value in enumerate(default_values):
+        for i, default_value in enumerate(default_locations):
             entry_row = entry_rows[i]
             entry_row.delete(0, tk.END)  # Clear the current content
             entry_row.insert(tk.END, default_value)  # Set the default value
@@ -349,16 +292,6 @@ def location_input_dialog_box():
             elif mode == "linear": select_linear(i)
             elif mode == "volume": select_volume(i)
 
-        # Set the horizontal and vertical entry boxes with default values
-        for i, default_value in enumerate(default_horizontal):
-            entry_horizontal = entry_horizontals[i]
-            entry_horizontal.delete(0, tk.END)
-            entry_horizontal.insert(tk.END, default_value)
-
-        for i, default_value in enumerate(default_vertical):
-            entry_vertical = entry_verticals[i]
-            entry_vertical.delete(0, tk.END)
-            entry_vertical.insert(tk.END, default_value)
 
         # Set the checkboxes in the fifth column based on default
         # Update: change peripheral folder checkbox to entry box
@@ -385,15 +318,118 @@ def location_input_dialog_box():
         crop_amount_entry.delete(0, tk.END)
         crop_amount_entry.insert(0, default_crop_amount)
 
+    def update_dropdown(new_selected_preset):
+        # Update the preset dropdown menu to reflect changes in preset_options_dict when a new preset is created or deleted
+        menu = preset_options_dropdown_menu['menu']
+        menu.delete(0, 'end')  # Clear the existing menu options
+
+        # Add updated options from preset_options_dict
+        for option in preset_options_dict.keys():
+            menu.add_command(
+                label=option,
+                command=lambda value=option: (selected_preset.set(value), set_new_preset(value))
+            )
+
+        # Update the displayed value
+        selected_preset.set(new_selected_preset)
+
+
+    def create_new_preset():
+        # Gets the current information
+        image_location_list = []
+        scan_modes = []
+        subfolder_entry = []
+
+        for i, entry_row in enumerate(entry_rows):
+            text = entry_row.get()
+            if text:
+                image_location_list.append(text)
+
+            if linear_vars[i].get():
+                scan_modes.append("linear")
+            elif radial_vars[i].get():
+                scan_modes.append("radial")
+            elif volume_vars[i].get():
+                scan_modes.append("volume")
+            else:
+                scan_modes.append("")
+
+            subfolder_entry.append(subfolder_vars[i].get())
+
+        od_before_os = od_os_checkbox_var.get()
+        subfolder_name = subfolder_name_entry.get()
+        unaveraged_images = unaveraged_checkbox_var.get()
+        min_contrast = min_contrast_var.get()
+        max_contrast = max_contrast_var.get()
+        crop_amount = crop_amount_var.get()
+
+        # Asks the user to give the preset a new name
+        new_preset_name = simpledialog.askstring("Preset", "New preset name:")
+
+        if new_preset_name:
+            new_preset_details = {
+                "default_locations": image_location_list,
+                "default_scan_type": scan_modes,
+                "default_subfolder": subfolder_entry,
+                "default_min_contrast": min_contrast,
+                "default_max_contrast": max_contrast,
+                "default_crop_amount": crop_amount,
+                "default_subfolder_name": subfolder_name,
+                "od_os_checkbox_boolean": od_before_os,
+                "imagej_checkbox_boolean": False,
+                "unaveraged_checkbox_boolean": unaveraged_images
+            }
+            preset_options_dict[new_preset_name] = new_preset_details
+            
+            # Read the preference txt file and replace the line
+            with open(preference_file, 'r') as file:
+                lines = file.readlines()
+
+            # Replace the preset_options_dict line
+            for i, line in enumerate(lines):
+                if line.strip().startswith("preset_options_dict"):
+                    lines[i] = f"preset_options_dict = {preset_options_dict}" + "\n"
+                    break
+
+            # Write the modified lines back to the file
+            with open(preference_file, 'w') as file:
+                file.writelines(lines)
+
+            # Update the dropdown menu with the new preset
+            update_dropdown(new_preset_name)
+
+            print(f"Created {new_preset_name}")
+
+    def delete_preset():
+        current_preset = selected_preset.get()
+        if current_preset != "Previous settings":
+            del preset_options_dict[current_preset]
+
+            # Read the preference txt file and replace the line
+            with open(preference_file, 'r') as file:
+                lines = file.readlines()
+
+            # Replace the preset_options_dict line
+            for i, line in enumerate(lines):
+                if line.strip().startswith("preset_options_dict"):
+                    lines[i] = f"preset_options_dict = {preset_options_dict}" + "\n"
+                    break
+
+            # Write the modified lines back to the file
+            with open(preference_file, 'w') as file:
+                file.writelines(lines)
+
+            # Update the dropdown menu with the new preset
+            update_dropdown("")
+
+            print(f"Deleted {current_preset}")
 
     def confirm(event=None):
-        global image_location_list, od_before_os, scan_modes, subfolder_entry, subfolder_name, horizontal_image_list, vertical_image_list, unaveraged_images, min_contrast, max_contrast, crop_amount
+        global image_location_list, od_before_os, scan_modes, subfolder_entry, subfolder_name, unaveraged_images, min_contrast, max_contrast, crop_amount
         
         image_location_list = []
         scan_modes = []
         subfolder_entry = []
-        horizontal_image_list = []
-        vertical_image_list = []
 
         for i, entry_row in enumerate(entry_rows):
             text = entry_row.get()
@@ -412,19 +448,6 @@ def location_input_dialog_box():
             subfolder_entry.append(subfolder_vars[i].get())
             # Update: change peripheral folder checkbox to entry box
 
-        for i, entry_horizontal in enumerate(entry_horizontals):
-            text = entry_horizontal.get()
-            if text:
-                horizontal_image_list.append(text)
-            else:
-                horizontal_image_list.append("")
-
-        for i, entry_vertical in enumerate(entry_verticals):
-            text = entry_vertical.get()
-            if text:
-                vertical_image_list.append(text)
-            else:
-                vertical_image_list.append("")
 
         od_before_os = od_os_checkbox_var.get()
         # Update: extra options dialog box - change line:
@@ -440,48 +463,47 @@ def location_input_dialog_box():
         window.destroy()
         window.master.destroy()
 
+        preset_options_dict["Previous settings"]["default_locations"] = image_location_list
+        preset_options_dict["Previous settings"]["default_scan_type"] = scan_modes
+        preset_options_dict["Previous settings"]["default_subfolder"] = subfolder_entry
+        preset_options_dict["Previous settings"]["default_min_contrast"] = min_contrast
+        preset_options_dict["Previous settings"]["default_max_contrast"] = max_contrast
+        preset_options_dict["Previous settings"]["default_crop_amount"] = crop_amount
+        preset_options_dict["Previous settings"]["default_subfolder_name"] = subfolder_name
+        preset_options_dict["Previous settings"]["od_os_checkbox_boolean"] = od_before_os
+        preset_options_dict["Previous settings"]["imagej_checkbox_boolean"] = False
+        preset_options_dict["Previous settings"]["unaveraged_checkbox_boolean"] = unaveraged_images
 
         # Removing variables that will be replaced
         with open(preference_file, 'r') as file:
             file_content = file.read()
         file_content = file_content.replace("\n\n\n", "\n") # Making sure we're not just adding a ton of extra lines over time
         variables_to_replace = [
+            "preset_options_dict",
             "image_location_list",
             "od_before_os",
             "scan_modes",
             "subfolder_entry",
             "subfolder_name",
-            "horizontal_image_list",
-            "vertical_image_list",
             "unaveraged_images",
             "min_contrast",
             "max_contrast",
             "crop_amount",
         ]
-        lines = [line for line in file_content.splitlines() if not any(line.startswith(variable) for variable in variables_to_replace)]
+        lines_not_to_be_replaced = [line for line in file_content.splitlines() if not any(line.startswith(variable) for variable in variables_to_replace)]
         if imageJ_location_reenter == True:
             # Update: extra options dialog box - probably move this to wherever it is
             shortened_lines = []
-            for i in lines:
+            for i in lines_not_to_be_replaced:
                 if not (i.startswith("screenshot_directory") or i.startswith("macro_location") or i.startswith("imagej_location")):
                     shortened_lines.append(i)
-            lines = shortened_lines
+            lines_not_to_be_replaced = shortened_lines
 
         # Saving the user's inputs for future uses
         with open(preference_file, 'w') as file:
-            file.write('\n'.join(lines))
-            file.write(f"\nimage_location_list = {image_location_list}")
-            file.write(f"\nod_before_os = {od_before_os}")
-            file.write(f"\nscan_modes = {scan_modes}")
-            # Update: change peripheral folder checkbox to entry box - next two lines
-            file.write(f"\nsubfolder_entry = {subfolder_entry}")
-            file.write(f"\nsubfolder_name = {subfolder_name}")
-            file.write(f"\nhorizontal_image_list = {horizontal_image_list}")
-            file.write(f"\nvertical_image_list = {vertical_image_list}")
-            file.write(f"\nunaveraged_images = {unaveraged_images}")
-            file.write(f"\nmin_contrast = {min_contrast}")
-            file.write(f"\nmax_contrast = {max_contrast}")
-            file.write(f"\ncrop_amount = {crop_amount}")
+            file.write('\n'.join(lines_not_to_be_replaced))
+            file.write(f"\npreset_options_dict = {preset_options_dict}")
+
 
 
     def on_dialog_close(event=None):
@@ -494,7 +516,8 @@ def location_input_dialog_box():
     window.title("Image settings")
     window.protocol("WM_DELETE_WINDOW", on_dialog_close)
 
-    instructions_label = tk.Label(window, text="Please enter the order that the images were taken in and what scan mode was used. If radial\nscans were used, record what the horizontal and vertical images represent. If this is not\nprovided for the radial images, then the image will not be processed. These will be used for the\nfile name.", justify='left')
+    instruction_text = "Please enter the order that the images were taken in and what scan mode was used. These will be used for the file name. If you enter 'delete' for any location those images will be deleted. If radial scans are used, in the location box separate out the different images with a comma and space (i.e. angle0, angle90). The 'linear' mode also works for annular scans."
+    instructions_label = tk.Label(window, text= instruction_text, wraplength=475, justify='left')
     instructions_label.grid(padx=10, sticky='w')
 
     dialog_frame = tk.Frame(window)
@@ -507,8 +530,6 @@ def location_input_dialog_box():
     checkbox_linears = []
     checkbox_radials = []
     checkbox_volumes = []
-    entry_horizontals = []
-    entry_verticals = []
     subfolder_vars = []
     checkbox_subfolders = []
 
@@ -522,12 +543,8 @@ def location_input_dialog_box():
     label_radial.grid(row=0, column=2)
     label_volume = tk.Label(dialog_frame, text="volume")
     label_volume.grid(row=0, column=3)
-    label_horizontal = tk.Label(dialog_frame, text="horizontal")
-    label_horizontal.grid(row=0, column=4, sticky='s')
-    label_vertical = tk.Label(dialog_frame, text="vertical")
-    label_vertical.grid(row=0, column=5, sticky='s')
     label_subfolder = tk.Label(dialog_frame, text="subfolder")
-    label_subfolder.grid(row=0, column=6)
+    label_subfolder.grid(row=0, column=4)
 
 
 
@@ -540,37 +557,35 @@ def location_input_dialog_box():
         entry_row = entry_rows[-1]
         entry_row.insert(tk.END, location)
 
-        # Access entry_horizontal and entry_vertical from the lists
-        entry_horizontal = entry_horizontals[-1]
-        entry_vertical = entry_verticals[-1]
-        entry_horizontal.insert(tk.END, horizontal_image_list[i])
-        entry_vertical.insert(tk.END, vertical_image_list[i])
-        if scan_modes[i] != "radial":
-            entry_horizontals[i].config(state='disabled')
-            entry_verticals[i].config(state='disabled')
-
-        row_num += 1
 
 
+
+    row_num += 1
     button_frame = tk.Frame(window)
-    button_frame.grid()
+    button_frame.grid(row=row_num)
     add_button = tk.Button(button_frame, text="+", command=add_entry_row)
     add_button.pack(side='right')
     remove_button = tk.Button(button_frame, text="-", command=remove_entry_row)
     remove_button.pack(side='left')
+    row_num += 1
+
 
     od_os_checkbox_var = tk.BooleanVar(value=od_before_os)
     od_os_checkbox = tk.Checkbutton(window, text="OD eyes were imaged before OS eyes", variable=od_os_checkbox_var)
-    od_os_checkbox.grid(padx=10, sticky='w')
+    od_os_checkbox.grid(row=row_num, padx=10, sticky='w')
+    row_num += 1
 
     # Update: custom defaults - make this part of a separate dialog box (or honestly make it a button)
     imagej_checkbox_var = tk.BooleanVar(value=False)
     imagej_checkbox = tk.Checkbutton(window, text="Re-enter ImageJ file location", variable=imagej_checkbox_var)
     imagej_checkbox.grid(padx=10, sticky='w')
+    row_num += 1
 
     unaveraged_checkbox_var = tk.BooleanVar(value=unaveraged_images)
     unaveraged_checkbox = tk.Checkbutton(window, text="Save a copy of the unaveraged images", variable=unaveraged_checkbox_var)
-    unaveraged_checkbox.grid(padx=10, sticky='w')
+    unaveraged_checkbox.grid(row=row_num, padx=10, sticky='w')
+    row_num += 1
+    
 
     # Getting user input on subfolder name
     # Update: change peripheral folder checkbox to entry box - replace with text that says how many subfolders will be made
@@ -612,16 +627,34 @@ def location_input_dialog_box():
     crop_unit_label = tk.Label(crop_amount_frame, text="µm")
     crop_unit_label.grid(row=0, column=2)
     row_num += 1
-    
 
-    # Creating "Restore defaults" and "Confirm" buttons at the bottom of the screen
+    # Creating preset setting dropdown menu
+    presets_frame = tk.Frame(window)
+    presets_frame.grid(row=row_num, column=0, sticky='w', padx=10)
+    preset_label = tk.Label(presets_frame, text = "Preset options: ")
+    preset_label.grid(row=0, column=0)
+    selected_preset = tk.StringVar(value="Previous settings")
+    preset_options_dropdown_menu = ttk.OptionMenu(
+        presets_frame,
+        selected_preset,
+        selected_preset.get(),
+        *preset_options_dict.keys(),
+        command=lambda _: set_new_preset(selected_preset.get())
+    )
+    preset_options_dropdown_menu.grid(row=0, column=1)
+    create_preset_button = tk.Button(presets_frame, text="Create preset", command=create_new_preset)
+    create_preset_button.grid(row=0, column=2)
+    delete_preset_button = tk.Button(presets_frame, text="Delete preset", command=delete_preset)
+    delete_preset_button.grid(row=0, column=3)
+
+
+    # Creating "Confirm" button at the bottom of the screen
     # Update: custom defaults - include button "Options"
     button_frame = tk.Frame(window)
     button_frame.grid()
-    restore_button = tk.Button(button_frame, text="Restore defaults", command=restore_defaults)
-    restore_button.pack(side='left', pady=10)
     confirm_button = tk.Button(button_frame, text="Confirm", command=confirm)
-    confirm_button.pack(side='right', pady=10, padx=10)
+    confirm_button.pack(pady=10, padx=10)
+
 
 
     window.bind("<Escape>", on_dialog_close)
@@ -632,13 +665,12 @@ def location_input_dialog_box():
     window.grab_set()
 
     # Center the window on the screen
-    window.after(10, lambda: center_dialog_box(window))
-    # Update: other. Do we have to have this delayed? ^
+    window.after(1, lambda: center_dialog_box(window))
 
     window.mainloop()
 
 # Starting up the initial dialog box
-location_input_dialog_box()
+first_dialog_box()
 
 
 print("image_location_list:", image_location_list)
@@ -648,8 +680,6 @@ print("scan_modes:", scan_modes)
 # Update: change peripheral folder checkbox to entry box - this should be a list, not a boolean
 print("subfolder_entry:", subfolder_entry)
 print("subfolder_name:", subfolder_name)
-print("horizontal_image_list:", horizontal_image_list)
-print("vertical_image_list:", vertical_image_list)
 print("min_contrast:", min_contrast)
 print("max_contrast:", max_contrast)
 print("crop_amount:", crop_amount)
@@ -659,7 +689,7 @@ def remove_underscores(list):       # This needs to be done because I count unde
     for i, item in enumerate(list):
         list[i] = item.replace("_", "underscore")
 
-underscore_list = [image_location_list, horizontal_image_list, vertical_image_list]
+underscore_list = [image_location_list] # Previously also had horizontal_image_list, vertical_image_list. Leaving as is in case need to clean any other list
 for list in underscore_list:
     remove_underscores(list)
 
@@ -669,10 +699,7 @@ images_to_put_into_subfolder = []
 for i, image_type in enumerate(image_location_list):
     if subfolder_entry[i] is True:
         images_to_put_into_subfolder.append(image_type)
-        if horizontal_image_list[i] != "":
-            images_to_put_into_subfolder.append(horizontal_image_list[i])
-        if vertical_image_list[i] != "":
-            images_to_put_into_subfolder.append(vertical_image_list[i])
+
 
 
 #SELECTING THE DIRECTORY
@@ -681,7 +708,7 @@ for i, image_type in enumerate(image_location_list):
 root = tk.Tk()
 root.withdraw()  # Hide the root window
 
-# Prompt the user to select a directory using a dialog
+# Prompt the user to select a directory
 image_directory = filedialog.askdirectory()
 
 # Check if the user selected a directory or canceled the dialog
@@ -716,22 +743,83 @@ with open(imagej_settings_file, "w") as file:
     file.write(f"{crop_amount}\n{min_contrast}\n{max_contrast}\n")
 
 
-#PUTTING FILE NAMES TOGETHER IN LIST
+# PUTTING FILE NAMES TOGETHER IN LIST
 # Retrieve all file names in the directory
 fileNames = [file for file in os.listdir(image_directory) if os.path.isfile(os.path.join(image_directory, file))]
 
 # Update: volume scan - removed <and name.split("_")[2] != "V">
-filteredFiles = [name for name in fileNames if name.endswith(".OCT") and "RegAvg" not in name]  #Removing any files that contain "RegAvg" and aren't .OCT
+filtered_files = [name for name in fileNames if name.endswith(".OCT") and "RegAvg" not in name]  #Removing any files that contain "RegAvg" and aren't .OCT
 
-#Checking to make sure all images are from the same date and experiment
-# Update: handle multiple experiments from one date
-experimentDate = filteredFiles[0].split("_")[0]
-for file_name in filteredFiles:
-    if file_name.split("_")[0] != experimentDate:
-        messagebox.showerror("Error", "You must only use images from one date.\n\nIf multiple experiments were done on that date, you can only process one experiment at a time.\n\nYour image files start with the date followed by a number (e.g. 2023-05-08-001). The number and date has to be the same for all files.")
+
+# Determining which experiment to convert from that day
+def user_selects_what_experiment_to_process(list_of_options):
+    def on_confirm(event=None):
+        nonlocal selected_option
+        selected_option = options_var.get()
+        window.destroy()
+    
+    def on_dialog_close(event=None):
+        nonlocal selected_option
+        selected_option = None
+        window.destroy()
+
+    window = tk.Toplevel()
+    window.title("Experiment Selection")
+    window.protocol("WM_DELETE_WINDOW", on_dialog_close)
+    window.bind("<Escape>", on_dialog_close)
+    window.bind("<Return>", on_confirm)
+    
+      
+    # Instructions label
+    instructions = tk.Label(window, text="Please select an experiment to process:")
+    instructions.pack(pady=10)
+    
+    # Variable to store the selected option
+    options_var = tk.StringVar(value=list_of_options[0])  # Default selection
+    selected_option = list_of_options[0]  # Initialize selected option
+            
+    # Create radio buttons for each option
+    for option in list_of_options:
+        rb = tk.Radiobutton(window, text=option, variable=options_var, value=option)
+        rb.pack(anchor="w")
+    
+    # Confirm button
+    confirm_button = ttk.Button(window, text="Confirm", command=on_confirm)
+    confirm_button.pack(pady=10)
+    
+
+    # Center the window on the screen
+    center_dialog_box(window)
+
+    window.wait_window()  # Wait for the window to close
+    
+    return selected_option
+
+# If there are several different experiments for the date, user needs to select which one to run
+experiment_list = []
+selected_experiment = filtered_files[0].split("_")[0]
+for file_name in filtered_files:
+    date_and_experiment = file_name.split("_")[0]
+    if date_and_experiment not in experiment_list:
+        experiment_list.append(date_and_experiment)
+
+if len(experiment_list) > 1:
+    selected_experiment = user_selects_what_experiment_to_process(experiment_list)
+    if selected_experiment is None:
+        print("No option selected. Shutting down program.")
         exit()
 
-#Extracting info from the file names
+
+    temp_list = []    
+    for file_name in filtered_files:
+        if file_name.split("_")[0] == selected_experiment:
+            temp_list.append(file_name)
+    filtered_files = temp_list
+
+
+
+
+# Extracting info from the file names
 annotated_list = [
     [
         file_name,
@@ -739,13 +827,13 @@ annotated_list = [
         file_name.split("_")[2],  # Extract the image type ("L", "R", or "V")
         int(file_name.split("_")[-1].split(".")[0])  # Extract the identifying number
     ]
-    for file_name in filteredFiles
+    for file_name in filtered_files
 ]
-#Format: [file name, eye, image type, identifying number]
+# Format: [file name, eye, image type, identifying number]
 for sublist in annotated_list:   #interpreting what the image type is
     if sublist[2] == "R":
         sublist[2] = "radial"
-    elif sublist[2] == "L":
+    elif sublist[2] == "L" or "A":
         sublist[2] = "linear"
     elif sublist[2] == "V":
         sublist[2] = "volume"
@@ -768,7 +856,6 @@ for sublist in annotated_list:
     previousEye = currentEye
 
 number_of_mice = mouseNumber
-print(number_of_mice)
 
 #Next we will identify the location based off of the image type and order it appears in
 previousEye = None
@@ -786,10 +873,11 @@ for sublist in annotated_list:
     previousEye = currentEye
 
 
-#Figuring out how many images each mouse/eye has
-#annotated_list format: [0 file name, 1 eye, 2 image type, 3 ID number, 4 arbitrary mouse number]
-#Making a new list to define each eye image
-#eye_images_list format: [0 mouse, 1 eye, 2 image locations]
+
+# Figuring out how many images each mouse/eye has
+# annotated_list format: [0 file name, 1 eye, 2 image type, 3 ID number, 4 arbitrary mouse number]
+# Making a new list to define each eye image
+# eye_images_list format: [0 mouse, 1 eye, 2 image locations]
 eye_images_list = []
 
 location_mapping = {image_location: image_location[0].upper() for image_location in image_location_list}
@@ -852,14 +940,16 @@ def create_mouse_number_dialog_box():
     dialog_frame = ttk.Frame(window)
     dialog_frame.pack(padx=10, pady=10)
 
-    ttk.Label(dialog_frame, text=" ").grid(row=1, column=0, padx=5)
-    ttk.Label(dialog_frame, text="Mouse").grid(row=1, column=1, padx=5)
-    ttk.Label(dialog_frame, text="OD").grid(row=1, column=2, padx=5)
-    ttk.Label(dialog_frame, text="OS").grid(row=1, column=3, padx=5)
+
 
     # Add instructions label
     instructions_label = ttk.Label(dialog_frame, text="OD and OS columns signify what images the\nprogram thinks you have. If wrong, click Edit\nto manually adjust.")
     instructions_label.grid(row=0, column=0, columnspan=4, padx=5, pady=5, sticky="w")
+
+    ttk.Label(dialog_frame, text=" ").grid(row=1, column=0, padx=5)
+    ttk.Label(dialog_frame, text="Mouse").grid(row=1, column=1, padx=5)
+    ttk.Label(dialog_frame, text="OD").grid(row=1, column=2, padx=5)
+    ttk.Label(dialog_frame, text="OS").grid(row=1, column=3, padx=5)
 
     definition_of_acronymns = "\n".join([f"{value} = {key}" for key, value in location_mapping.items()])
     instructions_label2 = ttk.Label(dialog_frame, text=definition_of_acronymns)
@@ -921,7 +1011,7 @@ for number in inputted_mouse_numbers:
     mouse_number_dict[i] = number
     i += 1
 
-#Revising the annotatedList to include the real mouse numbers (or blanks if that was what was provided)
+#Revising the annotated_list to include the real mouse numbers (or blanks if that was what was provided)
 for sublist in annotated_list:
     newNumber = mouse_number_dict.get(sublist[4], "dialogBoxError")
     sublist[4] = newNumber
@@ -1028,7 +1118,7 @@ def create_individual_dialog_box(annotated_list):
     window.master.attributes("-alpha", 0)  # Set window opacity to 0 to make it invisible
 
     # Center the window on the screen
-    window.after(10, lambda: center_dialog_box(window))
+    window.after(1, lambda: center_dialog_box(window))
 
     window.mainloop()
 
@@ -1048,6 +1138,8 @@ for sublist in annotated_list:
         annotated_list_volume_scans.append(sublist)
     else:
         annotated_list_image_scans.append(sublist)
+
+
 
 # Saving the annotated_list to a txt file in the image directory
 annotated_volume_text_file = os.path.join(image_directory, "Annotated_list_of_volume_scans.txt")
@@ -1141,7 +1233,7 @@ def is_imagej_logo_visible():
     logo_path = os.path.join(screenshot_directory, "ImageJ_logo.png")
 
     if os.path.exists(logo_path):
-        logo_location = pyautogui.locateOnScreen(logo_path)
+        logo_location = pyautogui.locateOnScreen(logo_path, confidence=confidence)
         return logo_location is not None
 
     return False
@@ -1169,7 +1261,7 @@ else:
 #Installing the needed macros
 
 def click_on_image_center(image_path):
-    image_location = pyautogui.locateOnScreen(image_path)
+    image_location = pyautogui.locateOnScreen(image_path, confidence=confidence)
     if image_location:
         center_x, center_y = pyautogui.center(image_location)
         pyautogui.click(center_x, center_y)
@@ -1208,8 +1300,11 @@ wait_to_appear(os.path.join(screenshot_directory, "select_the_location_of_the_im
 
 pyautogui.typewrite(image_directory, interval=0.01)
 pyautogui.press('enter')
-time.sleep(0.5)
+time.sleep(2)
 
+
+# Making sure ImageJ is the active program still
+is_imagej_running()
 
 # Press 'a' key to trigger the ImageJ macro that converts OCT images to TIF images
 print("Sending command to ImageJ to convert OCT files")
@@ -1217,7 +1312,6 @@ pyautogui.press('a')
 
 time.sleep(2)
 
-# Update: volume scan - update code below to also monitor volume scan folder
 
 
 # Now the program will monitor the conversion of OCT images and will proceed with the rest of the
@@ -1253,28 +1347,62 @@ time.sleep(3)   # Give ImageJ time to close out of the image files
 def rename_files(directory, annotated_list, is_volume_scan):    # This will convert the original file names into what I want the file names to be
     file_list = os.listdir(directory)
     file_count = len(file_list)
+    
+    # Determine the number of scans per image, if it isn't a volume scan
+    number_of_images_in_scan = {}
+    if not is_volume_scan:
+        for file in file_list:
+            base_name = file[:-8]   # Excludes the 4-digit image sequence number and ".tif"
+            if base_name in number_of_images_in_scan:
+                number_of_images_in_scan[base_name] += 1
+            else:
+                number_of_images_in_scan[base_name] = 1
+
+    # Function to determine which of the location names within a radial scan to apply to a given image
+    def determine_radial_scan_location_name(name_with_sequence_number, location):
+        base_name = name_with_sequence_number[:-4]
+        sequence_number = int(name_with_sequence_number[-4:])
+        number_of_images = number_of_images_in_scan[base_name]
+        
+        location_list = location.split(", ")
+
+        images_per_scan = number_of_images // len(location_list)
+        scan_index = sequence_number // images_per_scan  # integer division to find the scan index
+        specific_radial_image_location = location_list[scan_index]  # Get the corresponding scan name from the list
+
+        return specific_radial_image_location
+
+
 
     for file_name_with_extension in file_list:
-        file_name = os.path.splitext(file_name_with_extension)[0]
+        file_name_without_extension = os.path.splitext(file_name_with_extension)[0]
 
         if not is_volume_scan:
-            file_name = file_name[:-2]  # Remove the last 2 characters - the image number from the image sequence
+            file_name_with_sequence_number = file_name_without_extension
+            file_name_without_sequence_number = file_name_without_extension[:-4]  # Remove the last 4 characters - the image number from the image sequence
+        else:
+            file_name_without_sequence_number = file_name_without_extension   # Volume scans don't have the sequence number attached
         
         for sublist in annotated_list:
             annotated_file_name = sublist[0]
             annotated_list_name_without_extension = annotated_file_name[:-4]  # Remove the last 4 characters (".OCT")
 
-            if file_name == annotated_list_name_without_extension:
+            if file_name_without_sequence_number == annotated_list_name_without_extension:
                 eye = sublist[1]
                 location = sublist[5]
                 mouse_number = sublist[4]
+                scan_type = sublist[2]
 
-                if is_volume_scan:
+                if scan_type == 'radial':
+                    location = determine_radial_scan_location_name(file_name_with_sequence_number, location)
+
+                # Figuring out the last part of the renamed file name
+                if scan_type == 'volume':
                     micron_depth = int(crop_amount)
                     mm_depth = micron_depth / 1000
                     final_part_of_file_name = f"{mm_depth:.3f}".rstrip('0').rstrip('.') + "mmdepth.avi"     # Reports up to 3 decimal places in name
                 else:   # For non-volume scans we need to keep the last identifier numbers. We'll add "mmdepth" later when averaging
-                    final_part_of_file_name = file_name_with_extension[-6:]
+                    final_part_of_file_name = file_name_with_extension[-8:] # includes the 4 sequence numbers and the ".tif"
 
                 new_name = f"{mouse_number}_{eye}_{location}_{final_part_of_file_name}"
                 break
@@ -1285,119 +1413,19 @@ def rename_files(directory, annotated_list, is_volume_scan):    # This will conv
         old_path = os.path.join(directory, file_name_with_extension)
         new_path = os.path.join(directory, new_name)
 
-        # Rename the file
-        os.rename(old_path, new_path)
+        # Rename the file if the location isn't "delete"
+        if location == "delete":
+            os.remove(old_path)
+        else:
+            os.rename(old_path, new_path)
 
-    print(f"Renamed {file_count} files in {os.path.basename(directory)}")
-
+    print(f"Renamed and/or deleted {file_count} files in {os.path.basename(directory)}")
 
 
 # Renaming the individual sequence files and volume scan files
 individual_sequence_images_directory = os.path.join(image_directory, "individual_sequence_images") # This was created in the ImageJ macro
 rename_files(individual_sequence_images_directory, annotated_list_image_scans, False)    # This will convert the original file names into what I want the file names to be
 rename_files(volume_scan_directory, annotated_list_volume_scans, True)
-
-
-
-# Deleting the unneeded horizontal images
-# Update: more radial scans - pretty sure you need to change this section to be more robust
-def edit_and_delete_sequence_images(directory_path):
-    file_list = os.listdir(directory_path)
-
-    # Group files based on mouse number, eye, and location
-    groups = {}
-    for filename in file_list:
-        parts = filename.split("_")
-        mouse_number = parts[0]
-        eye = parts[1]
-        location = parts[2]
-
-        group_key = (mouse_number, eye, location)
-        if group_key not in groups:
-            groups[group_key] = []
-        groups[group_key].append(filename)
-
-
-    horizontal_only_radial_images = []
-    vertical_only_radial_images = []
-    both_radial_images = []
-    delete_radial_images = []
-
-    horizontal_key = {}
-    vertical_key = {}
-    both_key = {}
-
-    for i, location in enumerate(image_location_list):
-        if scan_modes[i] == "radial":
-            if horizontal_image_list[i] == "" and vertical_image_list[i] != "":
-                vertical_only_radial_images.append(location)
-                vertical_key[location] = vertical_image_list[i]
-            elif vertical_image_list[i] == "" and horizontal_image_list[i] != "":
-                horizontal_only_radial_images.append(location)
-                horizontal_key[location] = horizontal_image_list[i]
-            elif horizontal_image_list[i] != "" and vertical_image_list != "":
-                both_radial_images.append(location)
-                both_key[location] = [horizontal_image_list[i], vertical_image_list[i]]
-            elif horizontal_image_list[i] == "" and vertical_image_list[i] == "":
-                delete_radial_images.append(location)
-
-
-    # Delete unnecessary files within each group
-    for group_files in groups.values():
-        # Determine the midpoint for the current group
-        image_numbers = [int(file.split("_")[3].split(".")[0]) for file in group_files]
-        midpoint = (max(image_numbers) + 1) // 2
-
-        if len(group_files) > 0 and group_files[0].split("_")[2] in delete_radial_images:
-            for file in group_files:
-                file_path = os.path.join(directory_path, file)
-                os.remove(file_path)
-
-        # Delete files with image numbers less than the midpoint (deleting the horizontal images)
-        if len(group_files) > 0 and group_files[0].split("_")[2] in vertical_only_radial_images:
-            location = group_files[0].split("_")[2]
-            for file in group_files:
-                image_number = int(file.split("_")[3].split(".")[0])
-                if image_number < midpoint:
-                    file_path = os.path.join(directory_path, file)
-                    os.remove(file_path)
-                else:
-                    new_name = file.replace(location, vertical_key[location])
-                    new_path = os.path.join(directory_path, new_name)
-                    os.rename(os.path.join(directory_path, file), new_path)
-
-        # Delete files with image numbers greater than the midpoint (deleting the vertical images)
-            # This only matters for radial images so if the user wants this they should just take linear images instead of taking radial and then deleting half of the image
-        if len(group_files) > 0 and group_files[0].split("_")[2] in horizontal_only_radial_images:
-            location = group_files[0].split("_")[2]
-            print("horizontal:", midpoint)
-            for file in group_files:
-                image_number = int(file.split("_")[3].split(".")[0])
-                if image_number >= midpoint:
-                    file_path = os.path.join(directory_path, file)
-                    os.remove(file_path)
-                else:
-                    new_name = file.replace(location, horizontal_key[location])
-                    new_path = os.path.join(directory_path, new_name)
-                    os.rename(os.path.join(directory_path, file), new_path)
-
-        # Rename files if location is using both horizontal and vertical images
-        for location in both_radial_images:
-            if len(group_files) > 0 and group_files[0].split("_")[2] == location:
-                for i, file in enumerate(group_files):
-                    if i < midpoint:
-                        new_name = file.replace(location, both_key[location][0])
-                    else:
-                        new_name = file.replace(location, both_key[location][1])
-
-                    new_path = os.path.join(directory_path, new_name)
-                    os.rename(os.path.join(directory_path, file), new_path)
-
-
-# Initiating edit_and_delete_sequence_images()
-edit_and_delete_sequence_images(individual_sequence_images_directory)
-print("Deleted unneeded files and renamed radial files")
-
 
 
 
@@ -1460,9 +1488,9 @@ def averaging_images(directory_path, image_directory):
         oct_volume_averager_text_alternate = os.path.join(screenshot_directory, "OCT_volume_averager2.png")
         search_start_time = time.time()
         while time.time() - search_start_time < 20:
-            averager_text_location = pyautogui.locateOnScreen(oct_volume_averager_text)
+            averager_text_location = pyautogui.locateOnScreen(oct_volume_averager_text, confidence=confidence)
             if averager_text_location is None:
-                averager_text_location = pyautogui.locateOnScreen(oct_volume_averager_text_alternate)
+                averager_text_location = pyautogui.locateOnScreen(oct_volume_averager_text_alternate, confidence=confidence)
                 if averager_text_location is not None:
                     pyautogui.click(centerOfButton(averager_text_location))
 
@@ -1488,21 +1516,21 @@ def averaging_images(directory_path, image_directory):
 
             # Checking to make sure the correct settings are selected
             # Checking and potentially changing the read setting
-            read_setting_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_read_setting.png")) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_read_setting2.png"))
-            if read_setting_loation is None:
-                read_setting_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_read_setting.png")) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_read_setting2.png"))
-                if read_setting_loation is None:
+            read_setting_location = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_read_setting.png"), confidence=confidence) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_read_setting2.png"), confidence=confidence)
+            if read_setting_location is None:
+                read_setting_location = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_read_setting.png"), confidence=confidence) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_read_setting2.png"), confidence=confidence)
+                if read_setting_location is None:
                     show_error_message("The script is unable to identify the read settings properly in the OCT Volume Averager box. Please select the correct setting before clicking okay.")
-                if read_setting_loation is not None:
-                    click_location = read_setting_loation.left + 10, read_setting_loation.top + 37
+                if read_setting_location is not None:
+                    click_location = read_setting_location.left + 10, read_setting_location.top + 37
                     pyautogui.click(click_location)
                     time.sleep(5)
 
 
             # Checking and potentially changing the save setting
-            save_setting_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_save_setting.png")) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_save_setting2.png"))
+            save_setting_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_save_setting.png"), confidence=confidence) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_save_setting2.png"), confidence=confidence)
             if save_setting_loation is None:
-                save_setting_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_save_setting.png")) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_save_setting2.png"))
+                save_setting_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_save_setting.png"), confidence=confidence) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "incorrect_save_setting2.png"), confidence=confidence)
                 if save_setting_loation is None:
                     show_error_message("The script is unable to identify the save settings properly in the OCT Volume Averager box. Please select the correct setting before clicking okay.")
                 if save_setting_loation is not None:
@@ -1514,13 +1542,13 @@ def averaging_images(directory_path, image_directory):
             # Locating the "okay" button
             pyautogui.moveTo(screen_width/2, 0)
             time.sleep(0.1)
-            okay_button_location = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "ok_button.png")) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "ok_button2.png"))
+            okay_button_location = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "ok_button.png"), confidence=confidence) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "ok_button2.png"), confidence=confidence)
             if okay_button_location is None:
                 show_error_message("The script is unable to locate the 'okay' button, either because it is being obscured or the style of the button has changed. If the later is the case, please take a screenshot of the button and save it in the ImageJ_clicking_files folder as ok_button.png. The script will now shut down.")
 
 
             # Checking the Options box
-            options_box_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_options_box_settings.png")) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_options_box_settings2.png"))
+            options_box_loation = pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_options_box_settings.png"), confidence=confidence) or pyautogui.locateOnScreen(os.path.join(screenshot_directory, "correct_options_box_settings2.png"), confidence=confidence)
             if options_box_loation is None:
                 show_error_message("Incorrect settings are selected\nPlease make the following changes as needed:\n\nUncheck 'Inverted Image Stack (EDI)'\nUncheck 'Keep Intermediate .tiff Files'\nSet the 'Number of Pixels to Crop' to 0 for both boxes\n\nAlternatively, something is covering part of the OCT Volume Averager dialog box and the script is unable to complete its confirmation that the proper settings are selected. Please make sure the dialog box is not partially hidden.\n\nAnother possibility is the style of the box has changed. In which case you would need to take a screenshot of it and save it as correct_options_box_settings.png in ImageJ_clicking_files.\n\nClick the okay button in this dialog box when you have made the changes.")
 
@@ -1529,13 +1557,13 @@ def averaging_images(directory_path, image_directory):
             averager_text_right_coordinate = averager_text_location.left + averager_text_location.width
             averager_text_bottom_coordinate = averager_text_location.top + averager_text_location.height
 
-            text_box_1 = averager_text_right_coordinate, averager_text_bottom_coordinate + 60
+            text_box_1 = read_setting_location.left + read_setting_location.width, read_setting_location.top + read_setting_location.height + 5
             pyautogui.click(text_box_1)
             pyautogui.hotkey('ctrl', 'a')
             pyautogui.typewrite(images_to_be_averaged_directory, interval=0.01)
             time.sleep(0.1)
 
-            text_box_2 = text_box_1 = averager_text_right_coordinate, averager_text_bottom_coordinate + 180
+            text_box_2 = save_setting_loation.left + save_setting_loation.width, save_setting_loation.top + save_setting_loation.height + 5
             pyautogui.click(text_box_2)
             pyautogui.hotkey('ctrl', 'a')
             pyautogui.typewrite(averaged_images_directory, interval=0.01)
@@ -1574,7 +1602,7 @@ def averaging_images(directory_path, image_directory):
         mm_depth = micron_depth / 1000
         mmdepth_text = f"{mm_depth:.3f}".rstrip('0').rstrip('.') + "mmdepth"     # Reports up to 3 decimal places in name
 
-        file_name = group_files[0][:-6] + mmdepth_text + ".tif"
+        file_name = group_files[0][:-8] + mmdepth_text + ".tif"
         new_path = os.path.join(averaged_images_directory, file_name)
 
         time.sleep(0.5)
@@ -1669,6 +1697,55 @@ if annotated_list_image_scans and unaveraged_images is True:
     restore_underscores(unaveraged_images_directory)
 
 
+# Combining converted files into one subdirectory
+def combining_directories(experiment):
+    final_directory_name = f"{experiment} OCT images"
+    final_directory_path = os.path.join(image_directory, final_directory_name)
+
+    # Create the subdirectory if it doesn't exist
+    if not os.path.exists(final_directory_path):
+        os.makedirs(final_directory_path)
+    else:
+        delete_directory(final_directory_path)
+        os.makedirs(final_directory_path)
+
+    directories_to_move = ["averaged_images", "unaveraged_images", "volume_scans"]
+
+    # Move each directory into the new subdirectory
+    for directory in directories_to_move:
+        source_directory = os.path.join(image_directory, directory)
+        destination_directory = os.path.join(final_directory_path, directory)
+
+        # Check if the source directory exists
+        if os.path.exists(source_directory):
+            shutil.move(source_directory, destination_directory)
+
+        # Deleting if nothing is in the directory
+        if os.path.isdir(destination_directory) and not os.listdir(destination_directory):
+            os.rmdir(destination_directory)
+    
+    # If there is only one subdirectory, move all its contents out to final_directory_path and delete the subdirectory
+    final_subdirectories_list = os.listdir(final_directory_path)
+    number_of_subdirectories = len(final_subdirectories_list)
+    
+    if number_of_subdirectories == 1:
+        only_subdirectory = final_subdirectories_list[0]
+        only_subdirectory_path = os.path.join(final_directory_path, only_subdirectory)
+
+        # Move all contents of the only_subdirectory to final_directory_path
+        for item in os.listdir(only_subdirectory_path):
+            item_path = os.path.join(only_subdirectory_path, item)
+            destination_path = os.path.join(final_directory_path, item)
+
+            # Move the item (file or subdirectory) to final_directory_path
+            shutil.move(item_path, destination_path)
+
+        # Delete the now-empty subdirectory
+        os.rmdir(only_subdirectory_path)
+    
+
+combining_directories(selected_experiment)
+
 # Deleting the intermediate directories
 delete_directory(individual_sequence_images_directory)
 delete_directory(cropped_tif_images_directory)
@@ -1685,3 +1762,5 @@ if os.path.isdir(averaged_images_directory) and not os.listdir(averaged_images_d
     os.rmdir(averaged_images_directory)
 if os.path.isdir(volume_scan_directory) and not os.listdir(volume_scan_directory):
     os.rmdir(volume_scan_directory)
+
+print("Complete")
